+++ app/com/lynxanalytics/biggraph/frontend_operations/GraphComputationOperations.scala
@@ -47,14 +47,15 @@ class GraphComputationOperations(env: SparkFreeEnvironment) extends ProjectOpera
           // Implemented on Spark.
           "Harmonic", "Lin", "Average distance",
           // From NetworKit.
-          // TODO: These two don't work. "No match for overloaded function call"
-          // "Closeness (approximate)", "Betweenness (estimate)",
-          // TODO: Laplacian is meant to be used with weights.
-          // "Laplacian",
+          "Closeness (estimate)", "Betweenness (estimate)", "Laplacian",
           "Betweenness", "Eigenvector", "Harmonic Closeness", "Katz",
           "K-Path", "Sfigality").sorted.toList)),
       Choice("direction", "Direction",
         options = Direction.attrOptionsWithDefault("outgoing edges")),
+      Choice("weight", "Edge weight",
+        options = FEOption.list("No weighting") ++ project.edgeAttrList[Double]),
+      NonNegInt("samples", "Sample size",
+        default = 1000, group = "Advanced settings"),
       NonNegInt("maxDiameter", "Maximal diameter to check",
         default = 10, group = "Advanced settings"),
       NonNegInt("bits", "Precision", default = 8, group = "Advanced settings"))
@@ -67,9 +68,13 @@ class GraphComputationOperations(env: SparkFreeEnvironment) extends ProjectOpera
       val es = Direction(
         params("direction"),
         project.edgeBundle, reversed = true).edgeBundle
-      def nk(algo: String) = graph_operations.NetworKitComputeAttribute.run(algo, es)
+      val weight =
+        if (params("weight") == "No weighting") None
+        else Some(project.edgeAttributes(params("weight")).runtimeSafeCast[Double])
+      def nk(algo: String) = graph_operations.NetworKitComputeDoubleAttribute.run(
+        algo, es, Map("samples" -> params("samples").toInt), weight)
       val centrality: Attribute[Double] = algorithm match {
-        case "Closeness (approximate)" => nk("ApproxCloseness")
+        case "Closeness (estimate)" => nk("ApproxCloseness")
         case "Betweenness" => nk("Betweenness")
         case "Eigenvector" => nk("EigenvectorCentrality")
         case "Betweenness (estimate)" => nk("EstimateBetweenness")
@@ -94,7 +99,7 @@ class GraphComputationOperations(env: SparkFreeEnvironment) extends ProjectOpera
       val name = params("name")
       // A directed graph just gets turned into an undirected graph.
       // We can skip that and just build an undirected graph.
-      val core = graph_operations.NetworKitComputeAttribute.run(
+      val core = graph_operations.NetworKitComputeDoubleAttribute.run(
         "CoreDecomposition", project.edgeBundle, Map("directed" -> false))
       project.newVertexAttribute(name, core, help)
     }
@@ -131,6 +136,89 @@ class GraphComputationOperations(env: SparkFreeEnvironment) extends ProjectOpera
             "tolerance" -> params("tolerance").toDouble), weight)
       }
       project.newVertexAttribute(name, positions, help)
+    }
+  })
+
+  register("Compute diameter")(new ProjectTransformation(_) {
+    params ++= List(
+      Param("name", "Save as", defaultValue = "diameter"),
+      NonNegDouble("max_error", "Maximum relative error", defaultValue = "0.1"))
+    def enabled = project.hasEdgeBundle
+    def apply() = {
+      val maxError = params("max_error").toDouble
+      val result = graph_operations.NetworKitComputeScalar.run(
+        "Diameter", project.edgeBundle, Map("directed" -> false, "max_error" -> maxError))
+      val name = params("name")
+      if (maxError == 0) {
+        project.newScalar(name, result.scalar1)
+      } else {
+        project.newScalar(name + "_lower", result.scalar1)
+        project.newScalar(name + "_upper", result.scalar2)
+      }
+    }
+  })
+
+  register("Compute effective diameter")(new ProjectTransformation(_) {
+    params ++= List(
+      Param("name", "Save as", defaultValue = "effective diameter"),
+      NonNegDouble("ratio", "Ratio to cover", defaultValue = "0.9"),
+      Choice("algorithm", "Algorithm", options = FEOption.list("estimate", "exact")),
+      NonNegInt("bits", "Extra bits", default = 7, group = "Esimation settings"),
+      NonNegInt("approximations", "Number of parallel approximations", default = 64,
+        group = "Esimation settings"))
+    def enabled = project.hasEdgeBundle
+    def apply() = {
+      val opts = Map("directed" -> false, "ratio" -> params("ratio").toDouble)
+      val result = params("algorithm") match {
+        case "exact" => graph_operations.NetworKitComputeScalar.run(
+          "EffectiveDiameter", project.edgeBundle, opts)
+        case "estimate" => graph_operations.NetworKitComputeScalar.run(
+          "EffectiveDiameterApproximation", project.edgeBundle, opts ++ Map(
+            "bits" -> params("bits").toInt,
+            "approximations" -> params("approximations").toInt))
+      }
+      project.newScalar(params("name"), result.scalar1)
+    }
+  })
+
+  register("Compute assortativity")(new ProjectTransformation(_) {
+    params ++= List(
+      Param("name", "Save as", defaultValue = "assortativity"),
+      Choice("attribute", "Attribute",
+        options = FEOption.unset +: project.vertexAttrList[Double]))
+    def enabled = project.hasEdgeBundle
+    def apply() = {
+      val attrName = params("attribute")
+      if (attrName != FEOption.unset.id) {
+        val result = graph_operations.NetworKitComputeScalar.run(
+          "Assortativity", project.edgeBundle, Map(), None,
+          Some(project.vertexAttributes(attrName).runtimeSafeCast[Double]))
+        project.newScalar(params("name"), result.scalar1)
+      }
+    }
+  })
+
+  register("Find optimal spanning tree")(new ProjectTransformation(_) {
+    params ++= List(
+      Param("name", "Save as", defaultValue = "in_tree"),
+      Choice("weight", "Edge weight",
+        options = FEOption.list("Unit weight") ++ project.edgeAttrList[Double]),
+      Choice("optimize", "Optimize for",
+        options = FEOption.list("Maximal weight", "Minimal weight")),
+      RandomSeed("seed", "Random seed", context.box))
+    def enabled = project.hasEdgeBundle
+    def apply() = {
+      val weightName = params("weight")
+      val weight = if (weightName == "Unit weight") {
+        project.edgeBundle.const(1.0)
+      } else {
+        project.edgeAttributes(weightName).runtimeSafeCast[Double]
+      }
+      val attr = graph_operations.NetworKitComputeDoubleEdgeAttribute.run(
+        "RandomMaximumSpanningForest", project.edgeBundle, Map("seed" -> params("seed").toLong),
+        Some(if (params("optimize") == "Maximal weight") weight
+        else graph_operations.DeriveScala.negative(weight)))
+      project.edgeAttributes(params("name")) = attr
     }
   })