+++ sphynx/lynxkite-sphynx/networkit_community_detection.go
@@ -3,6 +3,8 @@ package main
 
 import (
 	"fmt"
+	"log"
+	"runtime/debug"
 
 	"github.com/lynxkite/lynxkite/sphynx/networkit"
 )
@@ -13,6 +15,7 @@ func init() {
 			defer func() {
 				if e := recover(); e != nil {
 					err = fmt.Errorf("%v", e)
+					log.Printf("%v\n%v", e, string(debug.Stack()))
 				}
 			}()
 			vs := ea.getVertexSet("vs")